requirements.txt
Django==2.0.3
django-redis
hiredis
pycrypto
mock
django-rq==0.9.1
rq==0.6.0
rq-scheduler==0.6.1
